Fix bug where 'atof' was used instead of 'atoi'
This operation obtains the entity index, which requires 'atoi'. compiler casted it to int, but ideally we just convert it to an int right away.

+++ r5dev/vstdlib/callback.cpp
@@ -1098,7 +1098,7 @@ void BHit_f(const CCommand& args)
 
 	if (bhit_abs_origin->GetBool())
 	{
-		const int iEnt = atof(args[2]);
+		const int iEnt = atoi(args[2]);
 		if (const IClientEntity* pEntity = g_pClientEntityList->GetClientEntity(iEnt))
 		{
 			g_pDebugOverlay->AddSphereOverlay( // Render a debug sphere at the client's predicted entity origin.
